requirements/test.txt
pytest==2.8.5
pytest-allure-adaptor==1.6.8
